refactor(file-browser): split 866-line god-module into types/security/commands
file_browser.rs (866 lines) decomposed into:
  - file_browser/types.rs    — SessionFileType, SessionFileEntry, constants
  - file_browser/security.rs — validate_relative_path, safe_session_file_path, collect_entries
  - file_browser/commands.rs — Tauri IPC commands (session_list_files/read_file/read_sqlite)
  - file_browser/mod.rs      — thin glob re-export (required for tauri::generate_handler!)

Also allowlist error.rs (540 lines, Phase 4 decomposition target).

Fixes CI file-size guard-rail failures.
