+++ crates/tracepilot-tauri-bindings/src/commands/file_browser/mod.rs
@@ -1,0 +1,13 @@
+//! Session file browser Tauri commands.
+//!
+//! Provides read-only, sandboxed access to files within a session's
+//! directory. All paths are validated and canonicalized to prevent
+//! traversal attacks.
+
+mod commands;
+mod security;
+mod types;
+
+// Glob re-exports are required so `tauri::generate_handler!` can resolve the
+// hidden `__cmd__*` items generated by `#[tauri::command]`.
+pub use commands::*;
